Rearranging signing code locations
